refactored the code (base -> core)
