--- docs/notebooks/simsettings.txt
@@ -1,37 +1,0 @@
-PATHWORK = /home/zawadzkim/projects/pySWAP/docs/notebooks
-PATHATM = .
-PATHCROP = /home/zawadzkim/projects/pySWAP/docs/notebooks
-PATHDRAIN = /home/zawadzkim/projects/pySWAP/docs/notebooks
-SWSCRE = 0
-SWERROR = 0
-TSTART = 2018-01-01
-TEND = 2018-01-02
-NPRINTDAY = 1
-SWMONTH = 1
-SWYRVAR = 0
-OUTDAT = 
-2018-01-01
-2018-01-01
-OUTFIL = result
-SWHEADER = 0
-SWWBA = 0
-SWEND = 0
-SWVAP = 0
-SWBAL = 0
-SWBLC = 0
-SWSBA = 0
-SWATE = 0
-SWBMA = 0
-SWDRF = 0
-SWSWB = 0
-SWINI = 0
-SWINC = 0
-SWCRP = 0
-SWSTR = 0
-SWIRG = 0
-SWCSV = 1
-INLIST_CSV = 'rain,evap'
-SWCSV_TZ = 0
-SWAFO = 0
-SWAUN = 0
-SWDISCRVERT = 0
