+++ pyswap/core/all_variables.json
