+++ pyswap/core/database/events.py
