+++ pyswap/core/exeptions.py
