+++ pyswap/core/result.py
