+++ pyswap/core/simsettings.py
