+++ pyswap/core/swpmodel.py
