+++ pyswap/core/utils.py
