+++ pyswap/extras/solutetransport.py
@@ -1,7 +1,7 @@
 """Solute transport."""
 
 from dataclasses import dataclass
-from pyswap.base.dtypes import Subsection
+from pyswap.core.dtypes import Subsection
 from pandas import DataFrame
